move argument parser to module
