+++ src/arguments/__init__.py
@@ -1,0 +1,1 @@
+from .arguments import argument_parser, read_args
